requirements.txt
asgiref==3.5.2
Django==4.1.1
django-simpleui==2022.7.29
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==2.5.0
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
